Move public actions into its own folder
