+++ src/actions/public/get-btc-finality-by-block-hash.ts
