+++ src/decorators/public-bitcoin-kit-actions.ts
@@ -1,5 +1,5 @@
 import { actionsToDecorator } from "./actions-to-decorator.js";
-import * as bitcoinKitActions from "../actions/bitcoin-kit.js";
+import * as bitcoinKitActions from "../actions/public/bitcoin-kit.js";
 
 export const hemiPublicBitcoinKitActions = () =>
   actionsToDecorator(bitcoinKitActions);
