+++ src/decorators/public-bitcoin-tunnel-manager-actions.ts
@@ -1,5 +1,5 @@
 import { actionsToDecorator } from "./actions-to-decorator.js";
-import * as bitcoinTunnelManagerActions from "../actions/bitcoin-tunnel-manager.js";
+import * as bitcoinTunnelManagerActions from "../actions/public/bitcoin-tunnel-manager.js";
 
 export const hemiPublicBitcoinTunnelManagerActions = () =>
   actionsToDecorator(bitcoinTunnelManagerActions);
