+++ src/decorators/public-bitcoin-vault-actions.ts
@@ -1,5 +1,5 @@
 import { actionsToDecorator } from "./actions-to-decorator.js";
-import * as bitcoinVaultActions from "../actions/bitcoin-vault.js";
+import * as bitcoinVaultActions from "../actions/public/bitcoin-vault.js";
 
 export const hemiPublicBitcoinVaultActions = () =>
   actionsToDecorator(bitcoinVaultActions);
