+++ src/decorators/public-op-node-actions.ts
@@ -1,4 +1,4 @@
 import { actionsToDecorator } from "./actions-to-decorator.js";
-import * as opNodeActions from "../actions/get-btc-finality-by-block-hash.js";
+import * as opNodeActions from "../actions/public/get-btc-finality-by-block-hash.js";
 
 export const hemiPublicOpNodeActions = () => actionsToDecorator(opNodeActions);
