+++ src/index.ts
@@ -1,10 +1,10 @@
 export { hemi } from "./chains/hemi.js";
 export { hemiSepolia } from "./chains/hemi-sepolia.js";
 
-export * from "./actions/bitcoin-kit.js";
-export * from "./actions/bitcoin-tunnel-manager.js";
-export * from "./actions/bitcoin-vault.js";
-export * from "./actions/get-btc-finality-by-block-hash.js";
+export * from "./actions/public/bitcoin-kit.js";
+export * from "./actions/public/bitcoin-tunnel-manager.js";
+export * from "./actions/public/bitcoin-vault.js";
+export * from "./actions/public/get-btc-finality-by-block-hash.js";
 
 export { hemiPublicBitcoinKitActions } from "./decorators/public-bitcoin-kit-actions.js";
 export { hemiPublicBitcoinTunnelManagerActions } from "./decorators/public-bitcoin-tunnel-manager-actions.js";
